src/requirements.txt
numpy==1.21.2
pandas==1.3.3
PyYAML==6.0
scikit_learn==1.0.1
